implemented login, tokens, env variables and other features

+++ routers/post.py
@@ -3,10 +3,12 @@ from sqlalchemy.orm import Session
 from database import get_db
 import models
 import schemas
-from typing import List
+import oauth2
+from typing import List, Optional
 
 router = APIRouter(
-  prefix="/posts"
+  prefix="/posts",
+  tags=["Post"]
 )
 
 @router.get("/latestpost", response_model=schemas.Post)
@@ -17,9 +19,9 @@ def get_latest_post(db: Session = Depends(get_db)):
                             detail="There are no posts")
     return post
     
-@router.get("/", response_model=List[schemas.Post])
-def get_posts(db: Session = Depends(get_db)):
-    posts = db.query(models.Post).all()
+@router.get("", response_model=List[schemas.Post])
+def get_posts(db: Session = Depends(get_db), limit: int = 10, skip: int = 0, search: Optional[str] = ""):
+    posts = db.query(models.Post).filter(models.Post.title.contains(search)).limit(limit).offset(skip).all()
     return posts 
 
 @router.get("/{id}", response_model=schemas.Post)
@@ -32,20 +34,23 @@ def get_post(id: int, db: Session = Depends(get_db)):
     return post 
     
 
-@router.post("posts", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
-def create_posts(post: schemas.PostBase, db: Session = Depends(get_db)):
-    new_post = models.Post(**post.model_dump())
+@router.post("", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
+def create_posts(post: schemas.PostBase, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+    print(current_user)
+    new_post = models.Post(owner_id=current_user.id, **post.model_dump())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
     return new_post
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id: int, db: Session = Depends(get_db)):
+def delete_post(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if post == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id {id} does not exist")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
     db.delete(instance=post)
     db.commit()
@@ -53,12 +58,15 @@ def delete_post(id: int, db: Session = Depends(get_db)):
     return Response(status_code=status.HTTP_204_NO_CONTENT)
     
 @router.put("/{id}", response_model=schemas.Post)
-def update_post(id: int, updated_post: schemas.PostBase, db: Session = Depends(get_db)):
+def update_post(id: int, updated_post: schemas.PostBase, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id {id} does not exist")
+        
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
         
     post_query.update(updated_post.model_dump())
     db.commit()
